requirements-dev.txt
dj-database-url==0.5.0
dj-static==0.0.6
django==2.1.8
python-decouple==3.1
pytz==2018.5
static3==0.7.0
virtualenv==16.0.0
django-prometheus==1.0.15
django-model-utils==3.1.2
pytest==4.2.0
pytest-cov==2.6.1
pytest-django==3.4.7
mixer==6.1.3
django-guardian==1.5.0
sentry-sdk==0.7.7
raven==6.10.0
djangorestframework==3.9.2
markdown==3.1
django-filter==2.1.0
django-taggit==1.1.0
django-taggit-labels==0.6.0
django-taggit-serializer==0.1.7
django-mptt==0.9.1
django-crispy-forms==1.7.2
selenium
django-debug-permissions==1.0.0
django-progressive-web-app==0.1.1